split up generic tests
